ladybug_be/requirements_solar.txt
# ========== FastAPI & Server ==========
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20
python-dotenv==1.1.1

# ========== Ladybug Core (základní) ==========
ladybug-core==0.44.25
ladybug-geometry==1.34.13

# ========== Ladybug Radiance (solární simulace) ==========
ladybug-radiance==0.4.2
honeybee-radiance==1.67.8

# ========== Honeybee Core (HBJSON) ==========
honeybee-core==1.59.5
honeybee-schema==1.58.5

# ========== Výpočetní knihovny ==========
numpy==2.3.3
matplotlib==3.10.6

# ========== Ostatní ==========
pydantic==2.11.9
typing-extensions==4.15.0

# ========== Původní závislosti ==========
annotated-types==0.7.0
anyio==4.11.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
h11==0.16.0
httptools==0.6.4
idna==3.10
kiwisolver==1.4.9
packaging==25.0
pillow==11.3.0
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
watchfiles==1.1.0
websockets==15.0.1
